+++ companion/src/firmwares/multiprotocols.h
@@ -37,6 +37,7 @@ class Multiprotocols
       int protocol;
       unsigned int maxSubtype;
       bool hasFailsafe;
+      bool disableChannelMap;
       QStringList protocols;
       QString optionsstr;
     };
@@ -44,6 +45,7 @@ class Multiprotocols
     struct MultiProtocolDefinition {
       const int protocol;
       const bool hasFailsafe;
+      const bool disableChannelMap;
       const QStringList subTypeStrings;
       const QString optionsstr;
 
@@ -59,6 +61,7 @@ class Multiprotocols
       MultiProtocolDefinition(const radio_mm_definition &rd) :
         protocol(rd.protocol),
         hasFailsafe(rd.hasFailsafe),
+        disableChannelMap(rd.disableChannelMap),
         subTypeStrings(rd.protocols),
         optionsstr(rd.optionsstr)
       {
